change main interface signature and implementations

+++ bytefuzz.go
@@ -1,6 +1,6 @@
 package bytefuzz
 
 type Interface interface {
-	Distance(text, target []byte) int
-	DistanceString(text, target string) int
+	Distance(text, target []byte) float64
+	DistanceString(text, target string) float64
 }
